refactored :)
